feat: 更改全金币通关和普通通关的音效

+++ Assets/Scripts/Models/ScoreModel.cs
@@ -15,7 +15,7 @@ namespace Assets.Scripts {
             this.passDuration = passDuration;
         }
 
-        public float GoldScore() {
+        public int GoldScore() {
             return 100 * gold;
         }
